:recycle: Component to signin or signout in burger or usermenu
